use horizontal variants for text logos
The square one is a disaster for plymouth and doesn't match archlinux's
actual text log either which is also horizontal. Copy and paste the
Horizontal_ColorFull.svg from our artwork repo here and delete the
physical width/height section so it is a nice size. Big thanks to Naomi
Calabretta <nyaomicalabretta@gmail.com> for pointing this out.